vimrc: add the :DisableStripws and :WriteWithoutStripws commands
When editing a file with existing whitespace at ends of lines, the
whitespace is stripped automatically when saving. This may be
undesirable, because the file may be committed to a version control system
and stripping the whitespace may cause many insignificant differences
between the index and the working tree.

This patch adds a :DisableStripws command that disables stripping
whitespace when saving the file and the :WriteWithoutStripws command
that can be used to disable stripping once.